backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
python-dotenv==1.0.1
pymongo==4.5.0
google-generativeai==0.4.1
requests==2.31.0
pyjwt==2.8.0
cryptography==42.0.8
boto3==1.34.36
selenium==4.23.1
webdriver-manager==4.0.2
